Adding new information to db from the algorithm, developing home event with new method calling (with departure time instead of arrival);

+++ Implementation/travlendarServer/src/main/java/com/travlendar/travlendarServer/logic/util/googleJsonSubClass/Coordinates.java
@@ -4,9 +4,9 @@ public class Coordinates {
     Float lat;
     Float lng;
 
-    public Coordinates(Double lat, Double lng) {
-        this.lat = lat.floatValue();
-        this.lng = lng.floatValue();
+    public Coordinates(Float lat, Float lng) {
+        this.lat = lat;
+        this.lng = lng;
     }
 
     public Coordinates() {
